Visualizing Sustainability- Story
The number of scenes in a storyboard for a data visualization analysis of the factors affecting the financial independence, will depend on the complexity of the analysis and the specific insights that are trying to be conveyed. 

A storyboard is a visual representation of the data analysis process and it breaks down the analysis into a series of steps or scenes.